+++ packages/react/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* `<CuriProvider>` can accept a new `router` prop.
+
 ## 1.0.0-beta.29
 
 * `<Prefetch when>` accepts array of match function names to call and its render-invoked `children()` function now receives a third argument: `error`.